перенес получение токена, прогресса и блокирование UI в AuthViewController для соответствия с дизайном и ТЗ
