+++ ImageFeed/Base.lproj/Main.storyboard
@@ -3,7 +3,7 @@
     <device id="retina6_12" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21678"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21679"/>
         <capability name="Named colors" minToolsVersion="9.0"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="System colors in document resources" minToolsVersion="11.0"/>
@@ -133,7 +133,7 @@
                             </imageView>
                         </subviews>
                         <viewLayoutGuide key="safeArea" id="LH7-YI-E8b"/>
-                        <color key="backgroundColor" name="ypBlack"/>
+                        <color key="backgroundColor" name="ypRed"/>
                         <constraints>
                             <constraint firstItem="SkD-lZ-6a6" firstAttribute="centerY" secondItem="LH7-YI-E8b" secondAttribute="centerY" id="FJh-MS-Rsc"/>
                             <constraint firstItem="SkD-lZ-6a6" firstAttribute="centerX" secondItem="LH7-YI-E8b" secondAttribute="centerX" id="HXO-Iy-7BG"/>
@@ -384,6 +384,9 @@
         <image name="sharing_button" width="50" height="50"/>
         <namedColor name="ypBlack">
             <color red="0.10196078431372549" green="0.10588235294117647" blue="0.13333333333333333" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+        </namedColor>
+        <namedColor name="ypRed">
+            <color red="0.96078431372549022" green="0.41960784313725491" blue="0.42352941176470588" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
         </namedColor>
         <namedColor name="ypWhite">
             <color red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
