+++ ImageFeed/Utils/UIBlockingProgressHUD.swift
@@ -1,0 +1,25 @@
+//
+//  UIBlockingProgressHUD.swift
+//  ImageFeed
+//
+//  Created by Тимур Танеев on 10.06.2023.
+//
+
+import UIKit
+import ProgressHUD
+
+final class UIBlockingProgressHUD {
+    private static var window: UIWindow? {
+        return UIApplication.shared.windows.first
+    }
+
+    static func show() {
+        window?.isUserInteractionEnabled = false
+        ProgressHUD.show()
+    }
+
+    static func dismiss() {
+        window?.isUserInteractionEnabled = true
+        ProgressHUD.dismiss()
+    }
+}
